factor out workloads and helper methods
